requirements.txt
# Dota 2 Agent 依赖

# 核心依赖
requests>=2.28.0
python-dotenv>=1.0.0

# MCP 相关
mcp>=1.0.0

# 数据分析和可视化
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
pillow>=9.0.0

scikit-learn>=1.2.0

# 向量检索
faiss-cpu

# hello_agents>=0.1.0

openai>=1.0.0
openviking

# Web UI
fastapi
uvicorn